+++ src/main/java/com/nhnacademy/marketgg/server/httprequest/prodcutInquiryPostRequest.http
@@ -1,30 +1,30 @@
-POST http://localhost:7080/shop/v1/products/3/inquiries HTTP/1.1
+POST http://localhost:7080/shop/v1/products/4/inquiries HTTP/1.1
 Content-Type: application/json
 
 {
-  "memberId" : 5,
-  "title" : "상품 문의",
-  "content" : "문의 내용",
+  "memberId" : 1,
+  "title" : "상품 문의 222",
+  "content" : "문의 내용 222",
   "isSecret" : 0
 }
 
 ###
 
-GET http://localhost:7080/shop/v1/products/3/inquiries
+GET http://localhost:7080/shop/v1/products/4/inquiries
 Content-Type: application/json
 
 ###
 
-GET http://localhost:7080/shop/v1/mygg/product-inquiries/5
+GET http://localhost:7080/shop/v1/members/1/product-inquiries
 Content-Type: application/json
 
 ###
 
-PUT http://localhost:7080/shop/v1/admin/products/3/inquiries/6
+PUT http://localhost:7080/shop/v1/admin/products/4/inquiries/15
 Content-Type: application/json
 
 {
-  "adminReply" : "상품 문의 답변"
+  "adminReply" : "상품 문의 답변 22222"
 }
 
 ###
